Modify: ReportRepository(Integer -> Long)

+++ src/main/java/com/pacs/scanviewer/SCV/domain/ReportRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReportRepository extends JpaRepository<Report, Integer> { }
+public interface ReportRepository extends JpaRepository<Report, Long> { }
